+++ opm/core/props/BlackoilPropertiesBasic.cpp
@@ -19,7 +19,7 @@
 
 #include "config.h"
 #include <opm/core/props/BlackoilPropertiesBasic.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <iostream>
