+++ opm/core/props/IncompPropertiesBasic.cpp
@@ -21,7 +21,7 @@
 
 #include "config.h"
 #include <opm/core/props/IncompPropertiesBasic.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <iostream>
