+++ opm/core/props/pvt/PvtPropertiesIncompFromDeck.cpp
@@ -21,7 +21,7 @@
 #include "config.h"
 #include <opm/core/props/pvt/PvtPropertiesIncompFromDeck.hpp>
 #include <opm/core/props/phaseUsageFromDeck.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/core/props/BlackoilPhases.hpp>
 #include <opm/parser/eclipse/Deck/DeckItem.hpp>
