+++ opm/core/props/rock/RockCompressibility.cpp
@@ -20,7 +20,7 @@
 #include "config.h"
 #include <opm/core/props/rock/RockCompressibility.hpp>
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/common/ErrorMacros.hpp>
 #include <opm/common/OpmLog/OpmLog.hpp>
 #include <opm/core/utility/linearInterpolation.hpp>
