+++ opm/core/simulator/SimulatorTimer.cpp
@@ -20,7 +20,7 @@
 #include "config.h"
 #include <opm/core/simulator/SimulatorTimer.hpp>
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <ostream>
 #include <numeric>
