+++ opm/core/simulator/TimeStepControl.cpp
@@ -28,7 +28,7 @@
 #include <iostream>
 
 #include <opm/common/ErrorMacros.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/core/simulator/TimeStepControl.hpp>
 
 namespace Opm
