+++ opm/core/simulator/initState_impl.hpp
@@ -27,7 +27,7 @@
 #include <opm/core/grid.h>
 #include <opm/core/grid/GridHelpers.hpp>
 #include <opm/core/utility/MonotCubicInterpolator.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 #include <opm/core/props/IncompPropertiesInterface.hpp>
 #include <opm/core/props/BlackoilPropertiesInterface.hpp>
 #include <opm/core/props/phaseUsageFromDeck.hpp>
