+++ tests/test_satfunc.cpp
@@ -34,7 +34,7 @@
 #include <opm/core/pressure/msmfem/partition.h>
 
 #include <opm/core/utility/parameters/ParameterGroup.hpp>
-#include <opm/core/utility/Units.hpp>
+#include <opm/parser/eclipse/Units/Units.hpp>
 
 #include <array>
 #include <iostream>
